+++ src/yyjson.h
@@ -272,6 +272,17 @@
 #   endif
 #endif
 
+/** compile-time constant check for compiler */
+#ifndef yyjson_constant_p
+#   if yyjson_has_builtin(__builtin_constant_p) || (YYJSON_GCC_VER >= 3)
+#       define YYJSON_HAS_CONSTANT_P 1
+#       define yyjson_constant_p(value) __builtin_constant_p(value)
+#   else
+#       define YYJSON_HAS_CONSTANT_P 0
+#       define yyjson_constant_p(value) 0
+#   endif
+#endif
+
 /** deprecate warning */
 #ifndef yyjson_deprecated
 #   if YYJSON_MSC_VER >= 1400
@@ -483,34 +494,57 @@ yyjson_api uint32_t yyjson_version(void);
  * JSON Types
  *============================================================================*/
 
-/** Type of JSON value (3 bit). */
+/** Type of a JSON value (3 bit). */
 typedef uint8_t yyjson_type;
+/** No type, invalid. */
 #define YYJSON_TYPE_NONE        ((uint8_t)0)        /* _____000 */
+/** Raw string type, no subtype. */
 #define YYJSON_TYPE_RAW         ((uint8_t)1)        /* _____001 */
+/** Null type: `null` literal, no subtype. */
 #define YYJSON_TYPE_NULL        ((uint8_t)2)        /* _____010 */
+/** Boolean type, subtype: TRUE, FALSE. */
 #define YYJSON_TYPE_BOOL        ((uint8_t)3)        /* _____011 */
+/** Number type, subtype: UINT, SINT, REAL. */
 #define YYJSON_TYPE_NUM         ((uint8_t)4)        /* _____100 */
+/** String type, subtype: NONE, NOESC. */
 #define YYJSON_TYPE_STR         ((uint8_t)5)        /* _____101 */
+/** Array type, no subtype. */
 #define YYJSON_TYPE_ARR         ((uint8_t)6)        /* _____110 */
+/** Object type, no subtype. */
 #define YYJSON_TYPE_OBJ         ((uint8_t)7)        /* _____111 */
 
-/** Subtype of JSON value (2 bit). */
+/** Subtype of a JSON value (2 bit). */
 typedef uint8_t yyjson_subtype;
+/** No subtype. */
 #define YYJSON_SUBTYPE_NONE     ((uint8_t)(0 << 3)) /* ___00___ */
+/** False subtype: `false` literal. */
 #define YYJSON_SUBTYPE_FALSE    ((uint8_t)(0 << 3)) /* ___00___ */
+/** True subtype: `true` literal. */
 #define YYJSON_SUBTYPE_TRUE     ((uint8_t)(1 << 3)) /* ___01___ */
+/** Unsigned integer subtype: `uint64_t`. */
 #define YYJSON_SUBTYPE_UINT     ((uint8_t)(0 << 3)) /* ___00___ */
+/** Signed integer subtype: `int64_t`. */
 #define YYJSON_SUBTYPE_SINT     ((uint8_t)(1 << 3)) /* ___01___ */
+/** Real number subtype: `double`. */
 #define YYJSON_SUBTYPE_REAL     ((uint8_t)(2 << 3)) /* ___10___ */
+/** String that do not need to be escaped for writing (internal use). */
+#define YYJSON_SUBTYPE_NOESC    ((uint8_t)(1 << 3)) /* ___01___ */
 
-/** Mask and bits of JSON value tag. */
+/** The mask used to extract the type of a JSON value. */
 #define YYJSON_TYPE_MASK        ((uint8_t)0x07)     /* _____111 */
+/** The number of bits used by the type. */
 #define YYJSON_TYPE_BIT         ((uint8_t)3)
+/** The mask used to extract the subtype of a JSON value. */
 #define YYJSON_SUBTYPE_MASK     ((uint8_t)0x18)     /* ___11___ */
+/** The number of bits used by the subtype. */
 #define YYJSON_SUBTYPE_BIT      ((uint8_t)2)
+/** The mask used to extract the reserved bits of a JSON value. */
 #define YYJSON_RESERVED_MASK    ((uint8_t)0xE0)     /* 111_____ */
+/** The number of reserved bits. */
 #define YYJSON_RESERVED_BIT     ((uint8_t)3)
+/** The mask used to extract the tag of a JSON value. */
 #define YYJSON_TAG_MASK         ((uint8_t)0xFF)     /* 11111111 */
+/** The number of bits used by the tag. */
 #define YYJSON_TAG_BIT          ((uint8_t)8)
 
 /** Padding size for JSON reader. */
@@ -4378,6 +4412,54 @@ struct yyjson_doc {
  * Unsafe JSON Value API (Implementation)
  *============================================================================*/
 
+/*
+ Whether the string does not need to be escaped for serialization.
+ This function is used to optimize the writing speed of small constant strings.
+ This function works only if the compiler can evaluate it at compile time.
+ 
+ Clang supports it since v8.0,
+    earlier versions do not support constant_p(strlen) and return false.
+ GCC supports it since at least v4.4,
+    earlier versions may compile it as run-time instructions.
+ ICC supports it since at least v16,
+    earlier versions are uncertain.
+ 
+ @param str The C string.
+ @param len The returnd value from strlen(str).
+ */
+yyjson_api_inline bool unsafe_yyjson_is_str_noesc(const char *str, size_t len) {
+#if YYJSON_HAS_CONSTANT_P && (YYJSON_GCC_VER == 0 || YYJSON_GCC_VER >= 4)
+    if (yyjson_constant_p(len) && len <= 32) {
+        /*
+         Same as the following loop:
+         
+         for (size_t i = 0; i < len; i++) {
+             char c = str[i];
+             if (c < ' ' || c > '~' || c == '"' || c == '\\') return false;
+         }
+         
+         GCC evaluates it at compile time only if the string length is within 17
+         and -O3 (which turns on the -fpeel-loops flag) is used.
+         So the loop is unrolled for GCC.
+         */
+#       define yyjson_repeat32_incr(x) \
+            x(0)  x(1)  x(2)  x(3)  x(4)  x(5)  x(6)  x(7)  \
+            x(8)  x(9)  x(10) x(11) x(12) x(13) x(14) x(15) \
+            x(16) x(17) x(18) x(19) x(20) x(21) x(22) x(23) \
+            x(24) x(25) x(26) x(27) x(28) x(29) x(30) x(31)
+#       define yyjson_check_char_noesc(i) \
+            if (i < len) { \
+                char c = str[i]; \
+                if (c < ' ' || c > '~' || c == '"' || c == '\\') return false; }
+        yyjson_repeat32_incr(yyjson_check_char_noesc)
+#       undef yyjson_repeat32_incr
+#       undef yyjson_check_char_noesc
+        return true;
+    }
+#endif
+    return false;
+}
+
 yyjson_api_inline yyjson_type unsafe_yyjson_get_type(void *val) {
     uint8_t tag = (uint8_t)((yyjson_val *)val)->tag;
     return (yyjson_type)(tag & YYJSON_TYPE_MASK);
@@ -4531,9 +4613,8 @@ yyjson_api_inline yyjson_val *unsafe_yyjson_get_next(yyjson_val *val) {
 
 yyjson_api_inline bool unsafe_yyjson_equals_strn(void *val, const char *str,
                                                  size_t len) {
-    uint64_t tag = ((uint64_t)len << YYJSON_TAG_BIT) | YYJSON_TYPE_STR;
-    return ((yyjson_val *)val)->tag == tag &&
-    memcmp(((yyjson_val *)val)->uni.str, str, len) == 0;
+    return unsafe_yyjson_get_len(val) == len &&
+           memcmp(((yyjson_val *)val)->uni.str, str, len) == 0;
 }
 
 yyjson_api_inline bool unsafe_yyjson_equals_str(void *val, const char *str) {
@@ -4597,8 +4678,11 @@ yyjson_api_inline void unsafe_yyjson_set_real(void *val, double num) {
 }
 
 yyjson_api_inline void unsafe_yyjson_set_str(void *val, const char *str) {
-    unsafe_yyjson_set_type(val, YYJSON_TYPE_STR, YYJSON_SUBTYPE_NONE);
-    unsafe_yyjson_set_len(val, strlen(str));
+    size_t len = strlen(str);
+    bool noesc = unsafe_yyjson_is_str_noesc(str, len);
+    yyjson_subtype sub = noesc ? YYJSON_SUBTYPE_NOESC : YYJSON_SUBTYPE_NONE;
+    unsafe_yyjson_set_type(val, YYJSON_TYPE_STR, sub);
+    unsafe_yyjson_set_len(val, len);
     ((yyjson_val *)val)->uni.str = str;
 }
 
@@ -4730,6 +4814,7 @@ yyjson_api_inline const char *yyjson_get_type_desc(yyjson_val *val) {
         case YYJSON_TYPE_RAW  | YYJSON_SUBTYPE_NONE:  return "raw";
         case YYJSON_TYPE_NULL | YYJSON_SUBTYPE_NONE:  return "null";
         case YYJSON_TYPE_STR  | YYJSON_SUBTYPE_NONE:  return "string";
+        case YYJSON_TYPE_STR  | YYJSON_SUBTYPE_NOESC: return "string";
         case YYJSON_TYPE_ARR  | YYJSON_SUBTYPE_NONE:  return "array";
         case YYJSON_TYPE_OBJ  | YYJSON_SUBTYPE_NONE:  return "object";
         case YYJSON_TYPE_BOOL | YYJSON_SUBTYPE_TRUE:  return "true";
@@ -4779,7 +4864,8 @@ yyjson_api_inline size_t yyjson_get_len(yyjson_val *val) {
 
 yyjson_api_inline bool yyjson_equals_str(yyjson_val *val, const char *str) {
     if (yyjson_likely(val && str)) {
-        return unsafe_yyjson_equals_str(val, str);
+        return unsafe_yyjson_is_str(val) &&
+               unsafe_yyjson_equals_str(val, str);
     }
     return false;
 }
@@ -4787,7 +4873,8 @@ yyjson_api_inline bool yyjson_equals_str(yyjson_val *val, const char *str) {
 yyjson_api_inline bool yyjson_equals_strn(yyjson_val *val, const char *str,
                                           size_t len) {
     if (yyjson_likely(val && str)) {
-        return unsafe_yyjson_equals_strn(val, str, len);
+        return unsafe_yyjson_is_str(val) &&
+               unsafe_yyjson_equals_strn(val, str, len);
     }
     return false;
 }
@@ -4964,15 +5051,11 @@ yyjson_api_inline yyjson_val *yyjson_obj_get(yyjson_val *obj,
 yyjson_api_inline yyjson_val *yyjson_obj_getn(yyjson_val *obj,
                                               const char *_key,
                                               size_t key_len) {
-    uint64_t tag = (((uint64_t)key_len) << YYJSON_TAG_BIT) | YYJSON_TYPE_STR;
     if (yyjson_likely(yyjson_is_obj(obj) && _key)) {
         size_t len = unsafe_yyjson_get_len(obj);
         yyjson_val *key = unsafe_yyjson_get_first(obj);
         while (len-- > 0) {
-            if (key->tag == tag &&
-                memcmp(key->uni.ptr, _key, key_len) == 0) {
-                return key + 1;
-            }
+            if (unsafe_yyjson_equals_strn(key, _key, key_len)) return key + 1;
             key = unsafe_yyjson_get_next(key + 1);
         }
     }
@@ -5040,8 +5123,7 @@ yyjson_api_inline yyjson_val *yyjson_obj_iter_getn(yyjson_obj_iter *iter,
         }
         while (idx++ < max) {
             yyjson_val *next = unsafe_yyjson_get_next(cur + 1);
-            if (unsafe_yyjson_get_len(cur) == key_len &&
-                memcmp(cur->uni.str, key, key_len) == 0) {
+            if (unsafe_yyjson_equals_strn(cur, key, key_len)) {
                 iter->idx = idx;
                 iter->cur = next;
                 return cur + 1;
@@ -5534,7 +5616,18 @@ yyjson_api_inline yyjson_mut_val *yyjson_mut_real(yyjson_mut_doc *doc,
 
 yyjson_api_inline yyjson_mut_val *yyjson_mut_str(yyjson_mut_doc *doc,
                                                  const char *str) {
-    if (yyjson_likely(str)) return yyjson_mut_strn(doc, str, strlen(str));
+    if (yyjson_likely(doc && str)) {
+        size_t len = strlen(str);
+        bool noesc = unsafe_yyjson_is_str_noesc(str, len);
+        yyjson_subtype sub = noesc ? YYJSON_SUBTYPE_NOESC : YYJSON_SUBTYPE_NONE;
+        yyjson_mut_val *val = unsafe_yyjson_mut_val(doc, 1);
+        if (yyjson_likely(val)) {
+            val->tag = ((uint64_t)len << YYJSON_TAG_BIT) |
+                        (uint64_t)(YYJSON_TYPE_STR | sub);
+            val->uni.str = str;
+            return val;
+        }
+    }
     return NULL;
 }
 
@@ -5554,7 +5647,19 @@ yyjson_api_inline yyjson_mut_val *yyjson_mut_strn(yyjson_mut_doc *doc,
 
 yyjson_api_inline yyjson_mut_val *yyjson_mut_strcpy(yyjson_mut_doc *doc,
                                                     const char *str) {
-    if (yyjson_likely(str)) return yyjson_mut_strncpy(doc, str, strlen(str));
+    if (yyjson_likely(doc && str)) {
+        size_t len = strlen(str);
+        bool noesc = unsafe_yyjson_is_str_noesc(str, len);
+        yyjson_subtype sub = noesc ? YYJSON_SUBTYPE_NOESC : YYJSON_SUBTYPE_NONE;
+        yyjson_mut_val *val = unsafe_yyjson_mut_val(doc, 1);
+        char *new_str = unsafe_yyjson_mut_strncpy(doc, str, len);
+        if (yyjson_likely(val && new_str)) {
+            val->tag = ((uint64_t)len << YYJSON_TAG_BIT) |
+                        (uint64_t)(YYJSON_TYPE_STR | sub);
+            val->uni.str = new_str;
+            return val;
+        }
+    }
     return NULL;
 }
 
@@ -6233,15 +6338,11 @@ yyjson_api_inline yyjson_mut_val *yyjson_mut_obj_get(yyjson_mut_val *obj,
 yyjson_api_inline yyjson_mut_val *yyjson_mut_obj_getn(yyjson_mut_val *obj,
                                                       const char *_key,
                                                       size_t key_len) {
-    uint64_t tag = (((uint64_t)key_len) << YYJSON_TAG_BIT) | YYJSON_TYPE_STR;
     size_t len = yyjson_mut_obj_size(obj);
     if (yyjson_likely(len && _key)) {
         yyjson_mut_val *key = ((yyjson_mut_val *)obj->uni.ptr)->next->next;
         while (len-- > 0) {
-            if (key->tag == tag &&
-                memcmp(key->uni.ptr, _key, key_len) == 0) {
-                return key->next;
-            }
+            if (unsafe_yyjson_equals_strn(key, _key, key_len)) return key->next;
             key = key->next->next;
         }
     }
@@ -6327,8 +6428,7 @@ yyjson_api_inline yyjson_mut_val *yyjson_mut_obj_iter_getn(
         while (idx++ < max) {
             pre = cur;
             cur = cur->next->next;
-            if (unsafe_yyjson_get_len(cur) == key_len &&
-                memcmp(cur->uni.str, key, key_len) == 0) {
+            if (unsafe_yyjson_equals_strn(cur, key, key_len)) {
                 iter->idx += idx;
                 if (iter->idx > max) iter->idx -= max + 1;
                 iter->pre = pre;
@@ -6444,7 +6544,7 @@ yyjson_api_inline void unsafe_yyjson_mut_obj_add(yyjson_mut_val *obj,
 }
 
 yyjson_api_inline yyjson_mut_val *unsafe_yyjson_mut_obj_remove(
-    yyjson_mut_val *obj, const char *key, size_t key_len, uint64_t key_tag) {
+    yyjson_mut_val *obj, const char *key, size_t key_len) {
     size_t obj_len = unsafe_yyjson_get_len(obj);
     if (obj_len) {
         yyjson_mut_val *pre_key = (yyjson_mut_val *)obj->uni.ptr;
@@ -6452,8 +6552,7 @@ yyjson_api_inline yyjson_mut_val *unsafe_yyjson_mut_obj_remove(
         yyjson_mut_val *removed_item = NULL;
         size_t i;
         for (i = 0; i < obj_len; i++) {
-            if (key_tag == cur_key->tag &&
-                memcmp(key, cur_key->uni.ptr, key_len) == 0) {
+            if (unsafe_yyjson_equals_strn(cur_key, key, key_len)) {
                 if (!removed_item) removed_item = cur_key->next;
                 cur_key = cur_key->next->next;
                 pre_key->next->next = cur_key;
@@ -6482,8 +6581,7 @@ yyjson_api_inline bool unsafe_yyjson_mut_obj_replace(yyjson_mut_val *obj,
         yyjson_mut_val *cur_key = pre_key->next->next;
         size_t i;
         for (i = 0; i < obj_len; i++) {
-            if (key->tag == cur_key->tag &&
-                memcmp(key->uni.str, cur_key->uni.ptr, key_len) == 0) {
+            if (unsafe_yyjson_equals_strn(cur_key, key->uni.str, key_len)) {
                 cur_key->next->tag = val->tag;
                 cur_key->next->uni.u64 = val->uni.u64;
                 return true;
@@ -6525,8 +6623,7 @@ yyjson_api_inline bool yyjson_mut_obj_put(yyjson_mut_val *obj,
     key_len = unsafe_yyjson_get_len(key);
     yyjson_mut_obj_iter_init(obj, &iter);
     while ((cur_key = yyjson_mut_obj_iter_next(&iter))) {
-        if (key->tag == cur_key->tag &&
-            memcmp(key->uni.str, cur_key->uni.ptr, key_len) == 0) {
+        if (unsafe_yyjson_equals_strn(cur_key, key->uni.str, key_len)) {
             if (!replaced && val) {
                 replaced = true;
                 val->next = cur_key->next->next;
@@ -6566,8 +6663,7 @@ yyjson_api_inline yyjson_mut_val *yyjson_mut_obj_remove(yyjson_mut_val *obj,
     yyjson_mut_val *key) {
     if (yyjson_likely(yyjson_mut_is_obj(obj) && yyjson_mut_is_str(key))) {
         return unsafe_yyjson_mut_obj_remove(obj, key->uni.str,
-                                            unsafe_yyjson_get_len(key),
-                                            key->tag);
+                                            unsafe_yyjson_get_len(key));
     }
     return NULL;
 }
@@ -6576,8 +6672,7 @@ yyjson_api_inline yyjson_mut_val *yyjson_mut_obj_remove_key(
     yyjson_mut_val *obj, const char *key) {
     if (yyjson_likely(yyjson_mut_is_obj(obj) && key)) {
         size_t key_len = strlen(key);
-        uint64_t tag = ((uint64_t)key_len << YYJSON_TAG_BIT) | YYJSON_TYPE_STR;
-        return unsafe_yyjson_mut_obj_remove(obj, key, key_len, tag);
+        return unsafe_yyjson_mut_obj_remove(obj, key, key_len);
     }
     return NULL;
 }
@@ -6585,8 +6680,7 @@ yyjson_api_inline yyjson_mut_val *yyjson_mut_obj_remove_key(
 yyjson_api_inline yyjson_mut_val *yyjson_mut_obj_remove_keyn(
     yyjson_mut_val *obj, const char *key, size_t key_len) {
     if (yyjson_likely(yyjson_mut_is_obj(obj) && key)) {
-        uint64_t tag = ((uint64_t)key_len << YYJSON_TAG_BIT) | YYJSON_TYPE_STR;
-        return unsafe_yyjson_mut_obj_remove(obj, key, key_len, tag);
+        return unsafe_yyjson_mut_obj_remove(obj, key, key_len);
     }
     return NULL;
 }
@@ -6631,7 +6725,10 @@ yyjson_api_inline bool yyjson_mut_obj_rotate(yyjson_mut_val *obj,
         if (yyjson_likely(key)) { \
             size_t len = unsafe_yyjson_get_len(obj); \
             yyjson_mut_val *val = key + 1; \
-            key->tag = YYJSON_TYPE_STR | YYJSON_SUBTYPE_NONE; \
+            size_t key_len = strlen(_key); \
+            bool noesc = unsafe_yyjson_is_str_noesc(_key, key_len); \
+            key->tag = YYJSON_TYPE_STR; \
+            key->tag |= noesc ? YYJSON_SUBTYPE_NOESC : YYJSON_SUBTYPE_NONE; \
             key->tag |= (uint64_t)strlen(_key) << YYJSON_TAG_BIT; \
             key->uni.str = _key; \
             func \
@@ -6720,7 +6817,10 @@ yyjson_api_inline bool yyjson_mut_obj_add_str(yyjson_mut_doc *doc,
                                               const char *_val) {
     if (yyjson_unlikely(!_val)) return false;
     yyjson_mut_obj_add_func({
+        size_t val_len = strlen(_val);
+        bool val_noesc = unsafe_yyjson_is_str_noesc(_val, val_len);
         val->tag = ((uint64_t)strlen(_val) << YYJSON_TAG_BIT) | YYJSON_TYPE_STR;
+        val->tag |= val_noesc ? YYJSON_SUBTYPE_NOESC : YYJSON_SUBTYPE_NONE;
         val->uni.str = _val;
     });
 }
@@ -6786,8 +6886,7 @@ yyjson_api_inline yyjson_mut_val *yyjson_mut_obj_remove_strn(
         yyjson_mut_val *val_removed = NULL;
         yyjson_mut_obj_iter_init(obj, &iter);
         while ((key = yyjson_mut_obj_iter_next(&iter)) != NULL) {
-            if (unsafe_yyjson_get_len(key) == _len &&
-                memcmp(key->uni.str, _key, _len) == 0) {
+            if (unsafe_yyjson_equals_strn(key, _key, _len)) {
                 if (!val_removed) val_removed = key->next;
                 yyjson_mut_obj_iter_remove(&iter);
             }